refactor(core): relocate webhook runtime helpers to rise-backend-core
Move the backend-agnostic runtime resolvers out of the K8s webhook into a
new `rise_backend_core::runtime` module so both deployment backends share one
source of truth: `resolve_runtime_containers`, `should_have_infrastructure`,
`resolve_deployment_env_vars`, and the `DEPLOYING_/PRE_PUSHED_TIMEOUT_MINUTES`
constants. `webhook.rs` re-exports them so existing call paths are unchanged.

`ResolvedDeploymentEnvVars` is redefined with backend-neutral fields
(`Vec<(String, String)>` + `BTreeMap<String, Vec<u8>>`) instead of k8s_openapi
`EnvVar`/`ByteString`. The K8s webhook adapts back to those types at its two
consumption sites; the Docker reconciler consumes the neutral form directly.
Adds a `rise-deployment-spec` dependency to core (no cycle).

Behavior-preserving. `K8s-only` `compute_service_owner_per_container` stays in
`webhook.rs`. Verified: backend build + clippy clean; core runtime, webhook
env-var/runtime, and 138 Docker controller unit tests pass.
