+++ crates/rise-backend-core/Cargo.toml
@@ -17,6 +17,7 @@ schema = ["dep:schemars"]
 anyhow = { workspace = true }
 async-trait = { workspace = true }
 chrono = { workspace = true }
+rise-deployment-spec = { path = "../rise-deployment-spec" }
 schemars = { workspace = true, optional = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
@@ -24,4 +25,8 @@ serde_json = { workspace = true }
 # models. This crate never opens a connection or runs a query — the
 # `DeploymentStore` implementation lives in `rise-deploy`.
 sqlx = { workspace = true }
+tracing = { workspace = true }
 uuid = { workspace = true }
+
+[dev-dependencies]
+tokio = { workspace = true }
